requirements.txt
audioread==3.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
filelock==3.13.1
fonttools==4.49.0
fsspec==2024.2.0
gTTS==2.5.1
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
lazy-loader==0.3
librosa==0.10.1
llvmlite==0.41.1
MarkupSafe==2.1.5
matplotlib==3.7.5
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.1
noisereduce==3.0.0
numba==0.58.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
openai-whisper==20231117
packaging==23.2
pillow==10.2.0
platformdirs==4.2.0
playsound==1.3.0
pooch==1.8.1
PyAudio==0.2.14
pycparser==2.21
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.1
sympy==1.12
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.0
tqdm==4.66.2
transformers==4.38.0
triton==2.2.0
typing-extensions==4.9.0
urllib3==2.2.1
zipp==3.17.0
